Checking _64BIT_ flags, adding ARMFIX
Checking all _64BIT_ flags and adding relevant ARMFIX comments around to
make sure we fix them at the same time. There are a number of different
ARMFIX comments, each one with an explanation of what has to be done to
make it work on ARM. Other comments were added for clarification of why
that particular _64BIT_ flag is not at all platform specific.