HPCC-10640 Testsocket segfault on startup on ARM
Segfaults when trying to log the fact that the RDTSC instruction is not
available, before any logging has been set up.

This commit fixes both the crash when logging from within the jdebug setup
code (by ensuring that logging is set up earlier), and the fact that the
system logs the failure to use rdtsc on non-intel systems.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>