HPCC-10664 - Hold locks to subfiles, release super
Ensure that any activity that lookups a file, holds the locks to
the subfiles and releases the lock to the super. This allows
other jobs/components to manipulate the super, once the job has
established which files it needs.
Hold list of distributed files used by activity in base class
and convert any supers to list of subs just before query starts.

This commit also contains some tyding up:
+ making virtual declarations more consistent
+ common up the setAccessed calls
+ remove a few redundant method definitions

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>