HPCC-10683 Fix WU connection leak in WsWorkunits and a warning in ESP log
When a new job is submitted, WsWorkunit creates connections to
dali to get information about the WU. Due to a bug, one of the
connections is not released. The same problem also occurs when
a ZAP Report is created. The bug is fixed here.

The WsWorkunits::WUUpdate method is used when a user wants to
modify a wu. The last step of the method is to retrieve/send
back the updated information about the WU. The information
includes possible thor log files for the WU. The WUUpdate method
is also called when a new job is submitted. A Warning line "
WARNING: Cannot find TargetClusterInfo for workunit ..." is
written to esp log when the method trys to retrieve thor log
information. This is because the WU is just created and no
cluster has been specified for the WU yet. This fix skips the
warning code if no cluster has been specified for a WU.