HPCC-10749 Search WU/Query with non-existence of property/attr
Using the existing code, WUQuery?State=unknown does not return "unknown"
WUs. The "unknown" WUs are WUs which do not have the @state or @state =
"". We do not suport any xpath syntax that allow you to test for the non
-existence of a property or attribute. In this fix, the code is added to
filter ECL WUs (and DFU WU, Query) with non-existence of a property or
attribute.