HPCC-10653 Problems in recreateCloneSource if dali not available
If not connected to local dali, use cache info only.

Exceptions in looking up the remote file no longer fatal, even in leagacy
mode.

Add a longer expiry for group lookup cache when the lookup results in an
exception, together with a control command (control:resetremotedalicache)
to clear the cached information.

Make sure that remote file resolutions are cached for use in lockdali mode.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>