HPCC-10688 EclWatch misreporting partially suspended queries
Queries that are suspended on some roxie nodes but not others where not
always reported as suspended.  Make control:queries check all nodes.

Also if some nodes are suspended and others are not, indicate that
fact on EclWatch so that users are more aware that results and apparent
availability can be affected.

Signed-off-by: Anthony Fishbeck <anthony.fishbeck@lexisnexis.com>