HPCC-10821 Fix roxie job not shown on Activity page
The existing code goes through each cluster in environment settings and
find out the jobs of the cluster using /status/servers/server/@name=cluster's
'AgentQueue' in the environment settings. That does not work for roxie
cluster because @name='RoxieServer', not ='roxie.roxie'.

In this fix, the existing data collection code for cluster status and jobs is
divided into three steps. The step 1 focuses on collecting data about cluster
and cluster status. This step collects cluster information for each target
cluster list in environment settings and finds out whether cluster queue
is in status/servers or not. In the step 2, the code goes through each
target cluster (and other servers including ECLagent) list in status/servers
and collect job information (both running jobs and queued jobs) of the
cluster (and server). The data for roxie jobs will be included in this step.
The information collected in the step 1 and 2 will be added into
ActivityResponse by the code of step 3.

Signed-off-by: Kevin Wang kevin.wang@lexisnexis.com