HPCC-10883 Show which logical file causes exception in WsDfu/DFUQuery
An exception may be thrown when quering logical files and the file info of
a file cannot be read (for example, ambiguous xpath "Attr"). To help user
know the problem, it will be better to include the logical file name
inside the exception. In this fix, a try/catch is added to insert the file
name into the exception message.

Signed-off-by: Kevin Wang kevin.wang@lexisnexis.com