HPCC-10678 Include Package Map in new ECLWatch
In this fix, Package Map UI is enabled in new ECLWatch with minor code
changes. The existing code calls the getSelections() to retrieve target
cluster settings as soon as PackageMapQueryWidget is loaded. That does
not work for new ECLWatch. The new code calls the getSelections() after
the widget is initialized for display. Similar change is also made for
PackageMapValidateWidget.