HPCC-10920 Return ClusterName in WsDFU.DFUInfo response
If a logical file is associated with only one cluster, the cluster
name is returned in WsDFU.DFUInfo response. If a logical file is
associated with multiple clusters, the cluster names are returned
inside a comma separated string in the response.

Based on review comments, I created a new structure to hold both
cluster name and file parts on that cluster. Add an array of the
structure into DFUInfo response. Fill in data into that array. In
DFUInfo response, 'depr' existing cluster and DFUFileParts array.

The 'Copy' variable inside the DFUPart structure indicates whether
the copy is primary copy (Copy=1) or not.

Signed-off-by: Kevin Wang kevin.wang@lexisnexis.com