HPCC-10854 Roxie does not retry file resolution once query fails to load
If a query fails to load (because of missing files, the query will stay
suspended until either Roxie is restarted or the query is deleted and
republished.

Add an option to always retry loading queries if they failed to load
correctly (defaults true, but can be disabled in case it causes
performance issues on clusters where there are a lot of suspended
queries).

Also add an option to control:reload to foce retrying of failed queries

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>