HPCC-10941 Add mechanism for finding queries that use a given file
Build a cache of files used by queries.  Update cache when DALI changes
in QuerySets, PackageMaps, or PackageSets are detected.

Add method to WsWorkunits to WUListQueriesUsingFile.

Create a report xslt for the result, and add a link to the the
context menu for roxie files listed on the browse dfu files page.

Signed-off-by: Anthony Fishbeck <anthony.fishbeck@lexisnexis.com>