HPCC-11063 Upgrade sasha WUiterate() to return NumWUs to WsWorkunits
1. Add a new sasha command action: SCA_LIST_WITH_MATCHING_COUNT. In
this action, a flag 'NumberOfIdsMatching' is allowed to send to sasha
WUiterate(). If the flag is set to true (default to false), the
WUiterate() will count the number of workunits which match a group of
search filters and return the number to the caller. 2. Revise
WsWorkunits doWUQueryFromArchive() to send the sasha command with
the flag (if a user requests) and return the number to the user. 3.
Refactor the doWUQueryFromArchive() to be better encapsulated.

Signed-off-by: Kevin Wang kevin.wang@lexisnexis.com