HPCC-11087 - Track subs swapped as part of swap/promote
SwapSuperFile and PromoteSuperFile were not tracking the subs
removed/added between supers in a transaction. As a consequence
if there were incompatibilities between old and new subs, then
the transaction mechanism spuriously detected the old+new as
incompatible and threw and error, even though they weren't in the
same super at the same time.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>