HPCC-11147 Call logSOAPMessage() to log SOAP request (hide password)
When ESP processing time for a SOAP call is longer than predifined value,
ESP logs the SOAP request in case the log is used for debug. If the
SOAP request contains user password, it should not be shown in ESP log.
In this fix, the logSOAPMessage() is used to replace the existing log
code to hide the password.

Signed-off-by: Kevin Wang kevin.wang@lexisnexis.com