HPCC-11161 Default to roxie on Package Map pages
In this fix, I default the Target to the first roxie (if exist) and
the process to the first process of the roxie (if exist) for both
the Package Maps and Validate Package Map pages.

Also fix the problem of 'not find createChildTabID()' when loading
Package Map Details page. It is fixed by inheriting the
_TabContainerWidget which has the function.

Signed-off-by: Kevin Wang kevin.wang@lexisnexis.com