HPCC-11175 Fix issues with cache of files in use info
Set flag for local files that are not in a package or dfs superfile, so
they won't be ignored as in use by query.  Also fix issue with
getTargetClusters not returning any entries when processName="", so
calling WUListQueriesUsingFile with no cluster or process will look
on all roxies.

Also fix install of QueriesUsingFiles.xslt.

Also add a backdoor way of viewing the entire files in use cache, to
make debugging easier.  To see cache, go to
http://ip:port/WsWorkunits/filesinuse

Signed-off-by: Anthony Fishbeck <anthony.fishbeck@lexisnexis.com>

Signed-off-by: Anthony Fishbeck <anthony.fishbeck@lexisnexis.com>