HPCC-11177 Return TargetCluster info after WsSMC.PauseQueue
1. Create a new WsSMC method: GetStatusServerInfo() which returns
information from /Status/Servers/Server for a specified target cluster
or HPCC server; 2. The same method is called to return TargetCluster
info after WsSMC.PauseQueue and WsSMC.ResumeQueue; 3. Upgrade code
to display HPCC server even if there is no workunit on the server. A
user may pause or resume the server even if no WU is runnning on the
server.

Signed-off-by: Kevin Wang kevin.wang@lexisnexis.com