HPCC-11304 Field translation options not working correctly
There are several mechanisms for controlling whether field translation should
be allowed.

 1. In the workunit, using #option (layoutTranslationEnabled,1)
 2. In the package file, using control:enableFieldTranslation=”1”
 3. In the RoxieTopology.xml file, using fieldTranslationEnabled=”1”

The global setting can be overridden temporarily (until the next time “service
hpcc-init start” is executed) using testsocket

In theory, the workunit setting should (if present) override the package file,
which should in turn (if present) override the global setting. If none are
present, field translation is disabled.

However, the workunit setting was always used, defaulting to false if not
present.

Also make the control: optional in the package file for back-compatibility

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>