HPCC-11394 Support latest version of Rcpp
Latest version of Rcpp (included in the Ubuntu 14.04 repository) uses some
"clever" code to dynamically fix up references into Rcpp.so which is loaded
by RInside at startup.

Unfortunately if you try to make this call from code which is itself in a
shared object (as we do), the symbols creayed for the inline stubs in
Rcpp/routines.h get resolved first.

I couldn't find a cleaner way of fixing this than to use macros to mangle
the names of the stubs so that they do not clash.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>