HPCC-9310 Results as spills
Inactive at the moment until codegen options changed.
This commits adds support for graph results used as spills.
The advantage is that that results are to memory and only spilled
on demand as memory pressure requires and it addressed an issue
where distribution as not maintained if a child query required
a 'result' that was a spill.

Stage #1

There are quite a few auxiliary needed to utility classed to
avoid circular references. So that CThorGraphResult objects
can continue to reference a representation of the activity and
their utility objects, without link counting the original
activity and therefore the graphs those activities belong to.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>