HPCC-10684 Ensure old tracked session is removed from new
If an old session was being handled in stopSession, when a new
session (on same IP:port) started, the new session tracked the
old, but the old did not correctly remove itself from the current
state when it had finished stopping.

The consequence was that when the new came to stop at a later
point, it still had the old session tracked and tried [again]
to stop it again and instead of new.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>