HPCC-11570 Fix incorrect '@' display when WsDFU deleting superfile
When a superfile is deleted, ESP should not return '@null' or '@' as
a part of the file name inside WsDFU/DFUArrayAction Response.

For new ECLWatch to identify which files are failed to be deleted and
which files are deleted, two result arrays (SuccessActions and
FailedActions) are added into the WsDFU/DFUArrayAction Response. (for
HPCC-11677).

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>