HPCC-11570 Fix incorrect '@' display when WsDFU deleting superfile
When a superfile is deleted, ESP should not return '@null' or '@' as
a part of the file name inside WsDFU/DFUArrayAction Response.

For new ECLWatch to identify which files are failed to be deleted and
which files are deleted, a result array (ActionResults) is added into
the WsDFU/DFUArrayAction Response. Each item inside the array has a
flag to indicate whether the file is failed to be deleted(for
HPCC-11677).

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>

HPCC-11570 Merge two arrays to one

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>