HPCC-11702 Fix ambiguous path error when pausing a thor queue
Ambiguous path error shows when trying to pause/resume a multi-thor
cluster. The error is due to the recent changes for collecting and
sending out cluster info after a queue is paused. This fix upgrades
the code to read the cluster info from cache.

An active WU may show under multiple ECLServers when the servers
have the same queue. The problem is fixed by this fix.

Also remove Activity request from creation function of
Activity Info cache so that it can be called from other methods.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>