HPCC-11830 Fix bad group name lookup for multi instance ecalgent
For a cluster with multiple eclagent instances defined, eclagent
looks up the matching group name, comparing it's ip with the
pubilshed group ip's, to lookup it looks for groupnames that
match configuration name with extensions '_<n>'. The hthor
lookup code was incorrectly starting with a base of 1 vs 2 for
the 2nd instance.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>