HPCC-11620 Return permission list for current user
This fix upgrades ws_access.AccountPermissions to return
permission list for corrent user. If the AccountName in
AccountPermissions request is not set, the UserID with
that request is used as the AccountName for retrieving
the permission list for the corrent user. The current
user does not have to be an admin.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>