HPCC-11932 Roxie attach causing Roxie cluster to load empty package
The onReconnect calls are supposed to reload from new information in dali, but
becuase of the isConnected being set AFTER the onReconnects, and the iniCache
happening before, the onReconnect was liable to reload from the empty cache
information.

The cache should not be cleared on reconnects (only needed on the initial
connect, and that can be handled more cleanly by calling loadCache in
writeCache). The isConnected needs to be set earlier - it was moved recently,
I believe just because it seemed like a good idea... Added a comment to
explain why it needs to be where it is.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>