HPCC-11958 Add Roxie State check to preflight
The similar logic in ecl-roxie is used by this fix to
retrieve and format information about roxie state (hash).
The information is displayed inside ECLwatch preflight
page (a new column 'Roxie State'). WsMachine thread pool
is used to improve the performance of retrieving the
information.

Three functions of sending control query to roxie are
moved from esp/clients/roxiecontrol to common/roxiecommlib
based on Tony's recommendation. One of those functions is
used by this fix. The others may be used in future.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>