HPCC-12079 Start Activity Cache timeout from cache creation completion
The WsSMC.Activity method checks the cache timeout to decide
whether it can use the Activity information cache or not. The
existing code starts the cache timeout at the beginning of
creating the cache. If the cache creation takes a longer time
than the timeout value, the second Activity request may have
to re-build the cache as soon as the first Activity request
finishes the cache creation. This fix sets the starting time
of the timeout at the end of the cache creation.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>