HPCC-10617 Add support for Cassandra
Initial checkin of plugin to support embedding of Cassandra (CQL) queries.

Because the Cassandra cpp library does not provide information about the
expected types for bound parameters, and yet is very fussy that they match
what it wanted, this required a change to the embed API to allow the
information to be taken from the ECL parameter types.

Note that the Cassandra cpp interface libraries are currently only available
in source form (and building them relies on other libraries that are not
available on most distros), so we build and ship them as part of the hpcc
platform build.

Also changes the representation of fields information so that we can use
default field values for NULL fields, and implements this functionality in
all SQL-style plugins.

Also fixes trivial memory leak in eclagent spotted while testing.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>