HPCC-12059 Need tool to create initial LDAP OUs and admin HPCC user
Currently admins are required to enter the LDAP Admin credentials in the Config
Manager which then gets flushed to the environment and config files. This is
insecure and they are only needed to create the initial OUs and an HPCC Admin
user. This pull request introduces a new "initldap" tool that should be run
once after installing HPCC and running config manager. The tools asks the user
to enter an LDAP Admin user/pw and uses them to create all the OUs and the initial
HPCC Admin that is specified in the config "systemUser" fields

Signed-off-by: William Whitehead <william.whitehead@lexisnexis.com>