HPCC-12224 eclagent.log sometimes shows activity not related to workunit
The code that extracts a particular workunit's eclagent logfile from the master
logfile uses the process id as the key. However, it scans the entire line looking
for that PID, and when by chance that pid string appears elsewhere (ie as a thread
id) then it incorrectly includes that line in the logfile. This code change
locates the actual PID column in the logfile entry and compares it to the
search PID

Signed-off-by: William Whitehead <william.whitehead@lexisnexis.com>