HPCC-12311 Remove extra calls to Dali for JobQueues
The existing WsSMC.Activity code retrieves JobQueue info
using the functions inside wujobq and dfuwu. The
createJobQueue() and copyItemsAndState() are used for an
ECL job queue. The queuedJobs() is used for a DFL job queue.
Each of those functions creates at least one connection to
Dali. Since an HPCC systrem contains multiple job queues, a
lot of the connections are created for one onActivity call.
In this fix, only one connection is used to Dali which gets
the /JobQueues tree and the tree has the JobQueue info for
all of the job queues.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>