HPCC-12338 If localThor adjust slave/master default memory
Previously defaults were for both to grab 75%, which is 'ok' on
a standard distro. with overcommit_memory set to 0.
In practice the master rarely uses that much, so on a single node
system it wasn't an issue, however if overcommit_memory is = 2
Then master or slaves will fail with an error reserving memory.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>