HPCC-12506 Avoid getGraphs and esp service grabbing data too early
A odd 'isValid()' was used in a couple of places to check that
a graph was 'valid' by probing into it's guts.. In doing so
callers, e.g. esp WUInfo service, would end up pullin all the
graph data (instead of leaving it to be lazy fetched when
required).
This changes removes the isValid() which was pointless anyway
and also changes the graph creation so that the xgmml is supplied
explitly at graph create time.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>