HPCC-12490 Add getGraphsMeta() to explicitly get the meta info
getGraphsMeta retrieves the workunits graph meta data (name,
type, label, state) without getting the graph data.

The state comes from the graph progress, all graphs being
requested now, use a single connection to pull the state from
the progress. Previously it connected to each graph progress
node independently in series. No graph progress data is pulled
in the process.

The create graph process is changes slightly, to explicitly
pass in the xgmml and an odd 'IConstWUGraph::isValid()' is
removed, which had the negative effect of pulling a lot of the
graph data early.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>