HPCC-12522 Do not fetch extra data for WU Details page
At the moment, depending on how you access a WU Details page,
e.g. via ECL WU Search page or via Browse ECL WUs page, the
amount of information retrieved from ESP varies. More info
are retrieved in the case of searching it than via browsing
it. Consequently the time taken to bring up the workunit page
is a lot slower via search (if the workunit is big) than it
is via browse.

This fix cleans the code to not retrieve extra data for the
WU Details page.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>