HPCC-11993 Add URL fetch for bundle install
ecl-bundle can now accept a url in place of a filename.

If the url ends in .git, it is assumed to be a git repository (fetched using
git clone), otherwise it is assumed to be the url of a file that can be
retrived via curl.

In both cases the file/repository is fetched to a temporary local location,
processed as a local file/directory would be, then removed.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>