HPCC-12534 Fix regression introduced in last commit for this issue
SDS didn't like 0 timeout on a connect.
changeMode was fine, and is used to establish the exclusive lock
for a file after it has looked it up in non-exclusive mode.
However, DFS also updates the FS to remove empty scopes, which
with the previous change was using a connect with a 0 timeout and
failing, due to a spurious check before it had tried.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>