HPCC-12669 Rename platform.h and all references to it
The file named system/include/platform.h is somewhat of a common name among
the many SDKs out there, and causes build errors when trying to coexist
with one or more of these SDKs. A better solution would be to rename it to
something less ambiguous like hpcc_platform.h or hpccplatform.h. This PR
renames that file and updates all references to it

Signed-off-by: William Whitehead <william.whitehead@lexisnexis.com>