HPCC-12681 Not show WU Scope if no security setting
In WsWorkunits service and legacy ECLWatch, check if
a system has set LDAP feature authentication for
WsWorkunits. If not, the ECL WU Details page
should not set/display WU Scope field because the
restricted scope is only valid if the system has set
LDAP feature authentication for WsWorkunits.

The change for new ECLWatch UI will be in a separate
commit.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>