HPCC-12633 Add some memory buffering to spilling variety
Add a memory pool of a few chunks to disk write too.
This substantially speeds up splitter that can spill, when the
arms are actually close to each other.

Also change QueueOf::dequeue(element) to return element.
It makes it a little cleaner when disposing of a link counted
item.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>