HPCC-12819 Crash in roxie when no library is deployed
An exception thrown during graph creation (there may be other places that can
throw an exception, but library not found is probably the most common) would
cause the graph not to be cleaned up prior to the destructor of the base
context. If it is left until then you get a pure virtual call error.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>