HPCC-12372 Add ESP support for reporting locks
A new WsSMC method called GetLocks is added to report
locks/clients. The method calls the new function
getDALocks() in dadiags.cpp to retrieve filtered locks.
The getDALocks() sends a request to dali by calling the
getDaliDiagnosticValue(mb) with a new id "filteredlocks".
A new CCovenSDSManager::getFilteredLocks() is added to
dasds.cpp to read lock info and serialize the info into
a memory buffer. Ater the info is sent back to ESP, the
getDALocks() deserialize the lock info into a new class
CDALockIterator. WsSMC.GetLocks sets up its response
from the CDALockIterator.

This method is only available for admin users.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>