HPCC-12087 Ensure cached files released by implicit transaction
A few fileservice operations on superfiles, effectively leaked
the locks to subfiles, which in turn could cause deadlock when
subsequent super ops. tried to manipulate those subfiles.

This fix, adds an implicit transaction block around the effected
operations, to ensure that on exit the cache entries are cleared.
The added implicit transaction will have no effect, if these
methods are called from a user transaction.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>