HPCC-11917 Full keyed joins fail or worse unless key was remote
If the key file was remote (not part of cluster the query is
running on), then instead of sending the key fetch request to
the node in the local cluster which has the fetch part locally,
it should access the fetch directly from the requesting slaves.
However, for that to work, each slave had to have the full set
of fetch parts and correctly map the fpos to the correct part.

That logic was misisng, this PR detects if fetch file is remote
and ensures all slaves have access to all parts.

Also, unless the master happened to be on the same IP as slave 1,
it incorrectly thought the key was remote and failed for the
reasons above.

The upshot of above, was that Full Keyed Joins were essentially
broken unless running on a thor cluster with slaves on same IP
as master.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>